+++ packages/app/src/index.ts
@@ -1,1 +1,2 @@
+export type { TypesLocales, TypesTimeZones } from '@zerodep/types';
 export * from '@zerodep/utils';