+++ packages/to/to.string/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/toString';
+export * from './toString';