+++ packages/to/to/src/index.spec.ts
@@ -1,12 +1,13 @@
 import * as converters from './index';
 
-describe('Guards barrel package', () => {
+describe('To barrel package', () => {
   it('should export specific packages and interfaces', () => {
     expect(Object.keys(converters).sort()).toEqual([
       'ZeroDepErrorTo',
       'toJSON',
       'toJSONHOF',
       'toString',
+      'toStringHOF',
     ]);
   });
 });