+++ packages/types/types/src/index.ts
@@ -1,0 +1,2 @@
+export type { TypesLocales } from '@zerodep/types.locales';
+export type { TypesTimeZones } from '@zerodep/types.timezones';