+++ packages/utils/CHANGELOG.md
@@ -10,11 +10,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Introduce the `is.json` utility
 - Introduce the `is.promise` utility
+- Introduce the `is.typedarray` utility
 - Introduce the `ZeroDepErrorTo` class
 - Introduce the `to.json` converter
 
 ### Changed
 
+- Enhance the `to.string` detection and formatting algorithm
 - Performance improvements to `is.*` algorithms
 - ZeroDepError `tax` value renamed to `category` for clarity